@@ -1,6 +1,6 @@
 {
     "name": "afeefa/api-resources-server",
-    "version": "0.1.52",
+    "version": "0.2.3",
     "type": "library",
     "license": "MIT",
     "authors": [
@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=8.1",
+        "php": ">=8.2",
         "psr/container": "^1.0 || ^2.0",
         "symfony/filesystem": "^6.1"
     },
@@ -26,7 +26,7 @@
     },
     "config": {
       "platform": {
-          "php": "8.1"
+          "php": "8.2"
       }
     }
 }
